Add missing endpoints for debug ports
This adds missing debug endpoints in Devfiles where a debug command is
declared and expected to be listening on the debug port (injected via
the DEBUG_PORT environment variable).
This is required in tools like odo, so the endpoint can be detected and
port-forwarded using "odo dev --debug".
Also, Che is already having such endpoint for debug as well. See
https://github.com/devspaces-samples/quarkus-quickstarts/blob/devspaces-3-rhel-8/devfile.yaml?rgh-link-date=2022-09-14T08%3A11%3A25Z#L12-L15

Signed-off-by: Armel Soro <asoro@redhat.com>