Make sure to run odo v3 tests for stacks that have no starter projects defined
This is done via the `odo dev --no-commands` command (introduced in odo v3.12.0 [1]),
to make sure that a Dev Session could actually be started properly.
This is to cover cases for example where the container has a terminating command
or ends up being restarted for some reason, or `odo` is unable to sync files.

[1] https://odo.dev/docs/command-reference/dev#running-with-no-commands

Signed-off-by: Armel Soro <asoro@redhat.com>